settings modal

+++ tipovacka-app/src/app/components/match/match.component.ts
@@ -6,6 +6,9 @@ import {DataService} from '../../services/data.service';
 import {TranslocoPipe} from '@ngneat/transloco';
 import {AsyncPipe, NgClass, NgIf} from '@angular/common';
 import {ResultButtonComponent} from '../result-button/result-button.component';
+import {first} from 'rxjs';
+import {DialogService} from '@ngneat/dialog';
+import {SignInAlertComponent} from '../sign-in-alert/sign-in-alert.component';
 
 @Component({
   selector: 'app-match',
@@ -29,23 +32,36 @@ export class MatchComponent {
   sendingVote: VoteResult | undefined = undefined;
   public authService = inject(AuthService);
   private dataService = inject(DataService);
+  private dialogService = inject(DialogService);
 
   chooseResult = (chosenResult: VoteResult) => {
-    this.sendingVote = chosenResult;
-    if (chosenResult === this.votes[this.data.id]?.result) {
-      this.dataService.deleteVote(this.data.id)
-        .then(
-          () => {
-            this.sendingVote = undefined;
-          }
-        )
-      return;
-    }
-    this.dataService.addVote(this.data.id, chosenResult)
-      .then(
-        () => {
-          this.sendingVote = undefined;
-        }
+    this.authService.isSignIn$
+      .pipe(
+        first()
       )
+      .subscribe({
+        next: (isSignIn) => {
+          if (!isSignIn) {
+            this.dialogService.open(SignInAlertComponent);
+            return;
+          }
+          this.sendingVote = chosenResult;
+          if (chosenResult === this.votes[this.data.id]?.result) {
+            this.dataService.deleteVote(this.data.id)
+              .then(
+                () => {
+                  this.sendingVote = undefined;
+                }
+              )
+            return;
+          }
+          this.dataService.addVote(this.data.id, chosenResult)
+            .then(
+              () => {
+                this.sendingVote = undefined;
+              }
+            )
+        }
+      })
   }
 }
